[fix] Avoid comparing loop variables against out-of-range values

+++ src/indexed_bzip2/bzip2.hpp
@@ -402,11 +402,11 @@ Block::readBlockHeader()
     for ( int j = 0; j < groupCount; j++ ) {
         // Read lengths
         std::array<uint8_t, MAX_SYMBOLS> length;
-        uint16_t hh = getBits<5>();
+        unsigned int hh = getBits<5>();
         for ( unsigned int i = 0; i < symCount; i++ ) {
             while ( true ) {
                 // !hh || hh > MAX_HUFCODE_BITS in one test.
-                if ( MAX_HUFCODE_BITS - 1 < (unsigned)hh - 1 ) {
+                if ( MAX_HUFCODE_BITS - 1 < hh - 1 ) {
                     std::stringstream msg;
                     msg << "[BZip2 block header]  start_huffman_length " << hh
                     << " is larger than " << MAX_HUFCODE_BITS << " or zero\n";
